[change] 投稿メモを保存後呼び出し元のアクティビティに戻るように変更

+++ app/src/main/java/com/example/product/restaurantmemo/ShopLog.kt
@@ -10,5 +10,5 @@ open class ShopLog(
         open var id: Long = 1,
         open var placeId: String = "",
         open var comment: String? = null,
-        open var numStars: Int = 0
+        open var numStars: Float = 0f
 ): RealmObject(){}
